requirements.txt
gymnasium>=0.29
mujoco>=3.1
stable-baselines3[extra]>=2.3
torch>=2.2
numpy>=1.24
warp-lang>=1.12.0
mujoco-warp>=3.6.0
